requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
Faker==11.3.0
fractal==0.3.0
importlib-metadata==4.8.3
iniconfig==1.1.1
lxml==4.7.1
mpmath==1.2.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pygame==2.1.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.1
sympy==1.9
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
zipp==3.6.0
pytest-sugar
ipython
